Switch to Vitest as default testing framework
- Remove Jest and all Jest-related dependencies
- Add Vitest as the default testing framework
- Update package.json scripts to use Vitest
- Create vitest.config.ts for Vitest configuration
- Add Jest compatibility layer in tests/setupVitest.ts
- Update GitHub Actions workflow to use Vitest
- Clean up unnecessary Jest-related files and test runners
- Keep only essential Vitest Cloudflare Workers test runner

This provides a modern testing setup with Vitest while maintaining
compatibility with existing test files that use Jest syntax.
