--- cloudflare-test-runner/wrangler.toml
@@ -1,7 +1,0 @@
-name = "nylas-test-runner"
-main = "test-runner.mjs"
-compatibility_date = "2024-09-23"
-compatibility_flags = ["nodejs_compat"]
-
-[env.test.vars]
-NODE_ENV = "test"
