+++ package.json
@@ -14,12 +14,9 @@
     "node": ">=16"
   },
   "scripts": {
-    "test": "jest",
-    "test:coverage": "npm run test -- --coverage",
-    "test:cloudflare": "node run-comprehensive-jest-cloudflare.mjs",
-    "test:cloudflare:vitest": "node run-vitest-cloudflare.mjs",
-    "test:cloudflare:all": "node run-all-tests-cloudflare.mjs",
-    "test:cloudflare:wrangler": "node run-jest-cloudflare.mjs",
+    "test": "vitest",
+    "test:coverage": "vitest --coverage",
+    "test:cloudflare": "node run-vitest-cloudflare.mjs",
     "lint": "eslint --ext .js,.ts -f visualstudio .",
     "lint:fix": "npm run lint -- --fix",
     "lint:ci": "npm run lint:fix -- --quiet",
@@ -54,7 +51,6 @@
   },
   "devDependencies": {
     "@babel/core": "^7.3.3",
-    "@types/jest": "^29.5.2",
     "@types/mime-types": "^2.1.2",
     "@types/node": "^22.15.21",
     "@types/uuid": "^8.3.4",
@@ -66,10 +62,8 @@
     "eslint-plugin-custom-rules": "^0.0.0",
     "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-prettier": "^3.0.1",
-    "jest": "^29.6.1",
     "jest-fetch-mock": "^3.0.3",
     "prettier": "^3.5.3",
-    "ts-jest": "^29.1.1",
     "typedoc": "^0.28.4",
     "typedoc-plugin-rename-defaults": "^0.7.3",
     "typescript": "^5.8.3",
